backend/requirements.txt
# Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
apscheduler==3.10.4
jieba==0.42.1
requests==2.31.0
joblib==1.3.2
lightgbm==4.1.0


# Optional packages - install separately if needed:
# pandas - requires C++ compiler, install via: pip install pandas
# scikit-learn - requires C++ compiler
# selenium - for web scraping
# playwright - for web scraping

# Optional: For production database
# pymysql==1.1.0
# psycopg2-binary==2.9.9

